Renamed Donut to donut to follow PEP8.
