+++ donut/config.py
@@ -1,0 +1,13 @@
+# Database URI
+db_host = "localhost"
+db_user = "root"
+db_pass = "bAgEl82"
+db_name = "donut"
+DB_URI = "mysql://{0}:{1}@{2}/{3}".format(db_user, db_pass, db_host, db_name)
+
+# Only used for testing.
+TEST_PORT = 5050
+DEBUG = True
+
+# Secret key
+SECRET_KEY = "CzHCZY4TV9QULXnvMBxOrTxvgURPh1ya"
