+++ donut/modules/scheduler/__init__.py
@@ -1,4 +1,4 @@
 from flask import Blueprint
 blueprint = Blueprint('scheduler', __name__, template_folder='templates', static_folder='static')
 
-import Donut.modules.scheduler.routes
+import donut.modules.scheduler.routes
