+++ donut/modules/scheduler/static/css/scheduler.css
