+++ donut/modules/scheduler/templates/scheduler.html
