+++ donut/templates/contact.html
@@ -1,0 +1,89 @@
+{% extends "layout.html" %}
+{% block body %}
+
+
+<div class="container" style="background-color: #eee">
+    <div id="page-header">
+        <h1 id="pagetitle" >Important Campus Contact Numbers</h1>
+    </div>
+    <div id="mainbody">
+
+        <div id='article'><a name="Administration"></a><h2> <span class="mw-headline">Administration</span></h2>
+            <a name="Deans"></a><h3> <span class="mw-headline">Deans</span></h3>
+            <table style="width: 800px;">
+                <tr>
+                    <td style="width: 180px;"><b>Name</b>
+                    </td><td style="width: 400px;"><b>Description</b>
+                    </td><td><b>Email</b>
+                    </td></tr>
+                <tr>
+                    <td>John Dabiri
+                    </td><td>Dean of Undergraduate Students
+                    </td><td><a href="mailto:jodabiri@caltech.edu" class="external text" title="mailto:jodabiri@caltech.edu" rel="nofollow">jodabiri@caltech.edu</a>
+                    </td></tr>
+                <tr>
+                    <td>Barbara C. Green
+                    </td><td>Associate Dean of Students
+                    </td><td><a href="mailto:barbarag@caltech.edu" class="external text" title="mailto:barbarag@caltech.edu" rel="nofollow">barbarag@caltech.edu</a>
+                    </td></tr>
+                <tr>
+                    <td>Lesley Nye
+                    </td><td>Associate Dean of Undergraduate Students
+                    </td><td><a href="mailto:lnye@caltech.edu" class="external text" title="mailto:lnye@caltech.edu" rel="nofollow">lnye@caltech.edu</a>
+                    </td></tr>
+                <tr>
+                    <td>Beth Larranaga
+                    </td><td>Office Manager
+                    </td><td><a href="mailto:rosel@caltech.edu" class="external text" title="mailto:rosel@caltech.edu" rel="nofollow">rosel@caltech.edu</a>
+                    </td></tr>
+                <tr>
+                    <td>Kim Quintana
+                    </td><td>Office Assistant
+                    </td><td><a href="mailto:kquintan@caltech.edu" class="external text" title="mailto:kquintan@caltech.edu" rel="nofollow">kquintan@caltech.edu</a>
+                    </td></tr>
+            </table>
+            <a name="Student_Life"></a><h3> <span class="mw-headline">Student Life</span></h3>
+            <table style="width: 800px;">
+                <tr>
+                    <td style="width: 180px;"><b>Name</b>
+                    </td><td style="width: 400px;"><b>Description</b>
+                    </td><td><b>Email</b>
+                    </td></tr>
+                <tr>
+                    <td>Tom Mannion
+                    </td><td>Senior Director, Student Activities and Programs
+                    </td><td><a href="mailto:mannion@caltech.edu" class="external text" title="mailto:mannion@caltech.edu" rel="nofollow">mannion@caltech.edu</a>
+                    </td></tr>
+                <tr>
+                    <td>Erik Snowberg
+                    </td><td>MOSH (Master of Student Houses)
+                    </td><td><a href="mailto:snowberg@caltech.edu" class="external text" title="mailto:snowberg@caltech.edu" rel="nofollow">snowberg@caltech.edu</a>
+                    </td></tr>
+                <tr>
+                    <td>Peter Daily
+                    </td><td>Assistant VP, Student Affairs, Housing, Dining
+                    </td><td><a href="mailto:pdaily@caltech.edu" class="external text" title="mailto:pdaily@caltech.edu" rel="nofollow">pdaily@caltech.edu</a>
+                    </td></tr>
+                <tr>
+                    <td>Jon Webster
+                    </td><td>Associate Director Dining Services
+                    </td><td><a href="mailto:jwebster@caltech.edu" class="external text" title="mailto:jwebster@caltech.edu" rel="nofollow">jwebster@caltech.edu</a>
+                    </td></tr>
+                <tr>
+                    <td>Denise Okamoto
+                    </td><td>Assistant Director of Events, Activities, and Programs
+                    </td><td><a href="mailto:denise@caltech.edu" class="external text" title="mailto:denise@caltech.edu" rel="nofollow">denise@caltech.edu</a>
+                    </td></tr>
+                <tr>
+                    <td>Sue Chiarchiaro
+                    </td><td>Assistant Director of Housing for Living-Learning Programs
+                    </td><td><a href="mailto:suec@caltech.edu" class="external text" title="mailto:suec@caltech.edu" rel="nofollow">suec@caltech.edu</a>
+                    </td></tr>
+            </table>
+
+        </div>
+    </div>
+
+</div>
+
+{% endblock %}
